Use unsigned ints instead of unsigned shorts to represent UTF8 font glyphs

+++ src/modules/font/Font.h
@@ -41,8 +41,8 @@ public:
 
 	virtual Rasterizer *newRasterizer(Data *data, int size) = 0;
 	virtual Rasterizer *newRasterizer(love::image::ImageData *data, std::string glyphs) = 0;
-	virtual Rasterizer *newRasterizer(love::image::ImageData *data, unsigned short *glyphs, int length) = 0;
-	virtual GlyphData *newGlyphData(Rasterizer *r, unsigned short glyph) = 0;
+	virtual Rasterizer *newRasterizer(love::image::ImageData *data, unsigned int *glyphs, int length) = 0;
+	virtual GlyphData *newGlyphData(Rasterizer *r, unsigned int glyph) = 0;
 
 	// Implement Module
 	virtual const char *getName() const = 0;
